requirements.txt
Django>=6.0.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
sendgrid-django
python-dotenv
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0